Game kinda renders.
Broke thump module into four sub modules.  Game, Wad, Render, Base.  The idea is to be able to test Wad all by stand-alone and to then layer on Renderer testing.  Game module uses the other three modules as Maven resources.

+++ base/src/main/java/thump/base/FixedPoint.java
@@ -22,7 +22,7 @@ public class FixedPoint {
 
     // Fixme. __USE_C_FIXED__ or something.
     public static int mul(int a, int b) {
-        return (int) ((a * (long) b) >> FRACBITS);
+        return (int) (((long)a * (long) b) >> FRACBITS);
     }
 
     //
